requirements.txt
pandas
beautifulsoup4
yamlloader
psycopg2-binary
webdriver-manager
requests
selenium
boto3
apache-airflow
python-dotenv
